chore: establish new folder structure, move flow-specific hooks, and delete dead code
