+++ ui/catalog/src/components/DeployFlow/DigitalAssistant/index.ts
@@ -1,2 +1,2 @@
-export { DeployFlow } from "./DeployFlow";
+export { DeployFlow } from "./DigitalAssistantDeployFlow";
 export type { DeployFlowProps, DeployFormData, ServiceConfig } from "./types";
