+++ ui/catalog/src/components/DeployFlow/DigitalAssistant/types/StepTwo.types.ts
