+++ ui/catalog/src/components/DeployFlow/DigitalAssistant/utils/StepTwo.utils.ts
