+++ ui/catalog/src/components/DeployFlow/Services/components/ServiceCredentialDisplay.tsx
