--- ui/catalog/src/components/DynamicField/index.ts
@@ -1,1 +1,0 @@
-export { DynamicField } from "./DynamicField";
