+++ ui/catalog/src/components/index.ts
@@ -10,9 +10,9 @@ export type {
 export { default as SolutionCard } from "./SolutionCard";
 export type { SolutionCardProps } from "./SolutionCard";
 export { default as DeployedServicesTable } from "./DeployedServicesTable";
-export { DeployFlow } from "./DeployFlow";
+export { DeployFlow } from "./DeployFlow/DigitalAssistant";
 export type {
   DeployFlowProps,
   DeployFormData,
   ServiceConfig,
-} from "./DeployFlow";
+} from "./DeployFlow/DigitalAssistant";
