+++ ui/catalog/src/pages/Services/Services.tsx
@@ -6,7 +6,7 @@ import ServiceCardSkeleton from "@/components/ServiceCard/ServiceCardSkeleton";
 import type { ServiceDetailData } from "@/components";
 import styles from "./Services.module.scss";
 import { DeployedServicesTable } from "@/components";
-import { ServicesDeployFlow } from "@/components/ServicesDeployFlow";
+import { ServicesDeployFlow } from "@/components/DeployFlow/Services";
 import DeploymentDetails from "@/components/DeploymentDetails";
 import { useServices } from "@/hooks/useServices";
 import { servicesReducer, initialState } from "./types";
